keep AgentActionRunningEvents in mpc.ts
